refactor(markdown): move internal render-options type out of public API (FEC-984)
ReactMarkdownRenderOptions is internal coordination plumbing shared between
the streaming and non-streaming render paths, but the barrel re-export chain
(src/index -> components -> markdown -> markdown.types) leaked it onto the
published package surface. Move it into a non-barreled markdown.internal-types
module tagged @internal so it is no longer reachable from the package root;
repoint the two internal importers. No behavior change.
